@@ -48,28 +48,30 @@ classifiers =
     Development Status :: 5 - Production/Stable
     Environment :: Web Environment
     Framework :: Django
-    Framework :: Django :: 1.11
-    Framework :: Django :: 2.0
-    Framework :: Django :: 2.1
-    Framework :: Django :: 2.2
-    Framework :: Django :: 3.0
-    Framework :: Django :: 3.1
+    Framework :: Django :: 3.2
+    Framework :: Django :: 4.2
+    Framework :: Django :: 5.0
+    Framework :: Django :: 5.1
+    Framework :: Django :: 5.2
+    Framework :: Django :: 6.0
     Intended Audience :: Developers
     License :: OSI Approved :: MIT License
     Operating System :: OS Independent
     Programming Language :: Python
     Programming Language :: Python :: 3 :: Only
-    Programming Language :: Python :: 3.5
-    Programming Language :: Python :: 3.6
-    Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.9
+    Programming Language :: Python :: 3.10
+    Programming Language :: Python :: 3.11
+    Programming Language :: Python :: 3.12
+    Programming Language :: Python :: 3.13
     Topic :: Internet :: WWW/HTTP
 
 [options]
-python_requires = >=3.5
+python_requires = >=3.8
 install_requires =
-    djangorestframework
-    django-filter>=2.0
+    djangorestframework>=3.12
+    django-filter>=21.1
 
 zip_safe = False
 include_package_data = True
